# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Masjid Member App
# during dev, should point to your local machine
host: localhost:10010
# basePath prefixes all resource paths 
basePath: /
# 
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /auth:
    x-swagger-router-controller: authentication
    post:
      description: Bearer token
      # used as the method name of the controller
      operationId: getToken
      parameters:
        - name: credentials
          description: Pass all Mandatory login credentials i.e. user_name, password
          in: body
          required: true
          schema:
            # a pointer to a definition
            $ref: "#/definitions/CredentialRequest"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/CredentialResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

  /member:
    # binds a127 app logic to a route
    x-swagger-router-controller: member
    get:
      description: Find all members
      # used as the method name of the controller
      operationId: getAllMembers
      parameters:
        - in: query
          name: offset
          type: integer
          description: The number of items to skip before starting to collect the result set.
          required: false
          default: 0
        - in: query
          name: limit
          type: integer
          description: The numbers of items to return.
          required: false
          default: 10
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/MemberResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      description: Add new member
      # used as the method name of the controller
      operationId: addNewMember
      parameters:
        - name: member
          description: Pass all Mandatory fields i.e. member_id, first_name, last_name, email, phone_number, address1, city, state, zip, status
          in: body
          required: false
          schema:
            # a pointer to a definition
            $ref: "#/definitions/createMemberInput"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/MemberInfoResponse"

        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    patch:
      description: update member information or status.
      # used as the method name of the controller
      operationId: updateMember
      parameters:
        - name: member
          description: Pass Mandatory fields i.e. member_id
          in: body
          required: false
          schema:
            # a pointer to a definition
            $ref: "#/definitions/UpdateMemberInput"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/MemberInfoResponse"

        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"        
  /member/{id}:
    # binds a127 app logic to a route
    x-swagger-router-controller: member
    get:
      description: Find a member by member Id
      # used as the method name of the controller
      operationId: getMemberInfo
      parameters:
        - name: id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/MemberInfoResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /searchmembers:
    # binds a127 app logic to a route
    x-swagger-router-controller: member
    post:
      description: Search Members
      # used as the method name of the controller
      operationId: findMembers
      parameters:
        - name: member
          description: Search with member_id, first_name, last_name, email, phone_number, address1, city, state, zip, status
          in: body
          required: false
          schema:
            # a pointer to a definition
            $ref: "#/definitions/searchMemberInput"

      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/MemberResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"

  /swagger:
    x-swagger-pipe: swagger_raw
# complex objects have schema definitions
definitions:
  MemberResponse:
    required:
      - error
      - members
    properties:
      error:
        type: boolean
      count:
        type: integer
      total_page:
        type: integer
      offset:
        type: integer
      limit:
        type: integer
      members:
        type: array
        items:
          type: object
          properties:
            member_id:
              type: string
              description: Masjid member Id
            first_name:
              type: string
              description: First Name of the user 
            last_name:
              type: string
              description: Last Name of the user
            email:
              type: ['string','null']
              description: Member's email id
            phone_number: 
              type: string
              description: Member's phone number
            address1:
              type: string
              description: Member's address1
            address2:
              type: ['string','null']
              description: Member's Optional second address
            city:
              type: string
              description: Member's city
            state:
              type: string
              description: Member's state
            zip:
              type: string
              description: Member's zip code
            subscription:
              type: integer
              description: Member's monthly subscription
            auto_withdrawal:
              type: integer
              description: Subscription auto withdrawal 1/0.
            gender:
              type: string
              description: Member's gender
            share_info:
              type: integer
              description: Member's monthly subscription
            status:
              type: integer
              description: Membership status,


  MemberInfoResponse:
    required:
      - error
      - member
    properties:
      error:
        type: boolean
      member:
          type: object
          properties:
            member_id:
              type: string
              description: Masjid member Id
            first_name:
              type: string
              description: First Name of the user 
            last_name:
              type: string
              description: Last Name of the user
            email:
              type: ['string','null']
              description: Member's email id
            phone_number: 
              type: string
              description: Member's phone number
            address1:
              type: string
              description: Member's address1
            address2:
              type: ['string','null']
              description: Member's Optional second address
            city:
              type: string
              description: Member's city
            state:
              type: string
              description: Member's state
            zip:
              type: string
              description: Member's zip code
            subscription:
              type: integer
              description: Member's monthly subscription
            auto_withdrawal:
              type: integer
              description: Subscription auto withdrawal 1/0.
            gender:
              type: string
              description: Member's gender
            share_info:
              type: integer
              description: Member's monthly subscription
            status:
              type: integer
              description: Membership status

  searchMemberInput:
    type: object
    properties:
      member_id:
        type: string
        description: Masjid member Id
      first_name:
        type: string
        description: First Name of the user
      last_name:
        type: string
        description: Last Name of the user
      email:
        type: string
        description: Member's email id
      phone_number: 
        type: string
        description: Member's phone number
      address1:
        type: string
        description: Member's address1
      city:
        type: string
        description: Member's city
      state:
        type: string
        description: Member's state
      zip:
        type: string
        description: Member's zip code
      status:
        type: integer
        description: Membership status
  createMemberInput:
    type: object
    properties:
      member_id:
        type: string
        description: Masjid member Id
      first_name:
        type: string
        description: First Name of the user
      last_name:
        type: string
        description: Last Name of the user
      email:
        type: ['string','null']
        description: Member's email id
      phone_number: 
        type: string
        description: Member's phone number
      address1:
        type: string
        description: Member's address1
      address2:
        type: ['string','null']
        description: Member's address2
      city:
        type: string
        description: Member's city
      state:
        type: string
        description: Member's state
      zip:
        type: string
        description: Member's zip code
      subscription:
        type: number
        description: Subscription amount per month
      auto_withdrawal:
        type: integer
        description: Subscription auto withdrawal 1/0.
      gender:
        type: string
        description: Member's gender M/F.
      share_info:
        type: integer
        description: Does Member want to share his/her info 1/0.
      status:
        type: integer
        description: Membership status
    required:
      - member_id
      - first_name
      - last_name
      - phone_number
      - address1
      - city
      - state
      - zip
      - subscription
      - gender
      - share_info
  UpdateMemberInput:
    type: object
    properties:
      member_id:
        type: string
        description: Masjid member Id
      first_name:
        type: string
        description: First Name of the user
      last_name:
        type: string
        description: Last Name of the user
      email:
        type: ['string']
        description: Member's email id
      phone_number: 
        type: string
        description: Member's phone number
      address1:
        type: string
        description: Member's address1
      address2:
        type: ['string']
        description: Member's address2
      city:
        type: string
        description: Member's city
      state:
        type: string
        description: Member's state
      zip:
        type: string
        description: Member's zip code
      subscription:
        type: number
        description: Subscription amount per month
      auto_withdrawal:
        type: integer
        description: Subscription auto withdrawal 1/0.
      gender:
        type: string
        description: Member's gender M/F.
      share_info:
        type: integer
        description: Does Member want to share his/her info 1/0.
      status:
        type: integer
        description: Membership status
    required:
      - member_id
  
  CredentialRequest:
    type: object
    properties:
      user_name:
        type: string
        description: Valid username
      password:
        type: string
        description: valid password
    required:
      - user_name
      - password
  
  CredentialResponse:
    properties:
      error:
        type: boolean
      token:
        type: string
        description: JWT token
    required:
      - error
      - token

  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string